backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.15.1",
    "bcrypt>=4.3.0",
    "email-validator>=2.2.0",
    "fastapi[standard]>=0.115.11",
    "jinja2>=3.1.6",
    "nanoid>=2.0.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary]>=3.2.5",
    "pydantic-settings>=2.8.1",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.38",
]
